gate_rag_system/requirements.txt
streamlit
faiss-cpu
sentence-transformers
groq
pymupdf
numpy
requests